cast option values to proper type
clang is more strict than gcc

+++ cb/allBalances.cpp
@@ -134,10 +134,10 @@ struct AllBalances:public Callback
         allAddrs.reserve(15 * 1000 * 1000);
 
         optparse::Values &values = parser.parse_args(argc, argv);
-        cutoffBlock = values.get("atBlock");
-        showAddr = values.get("withAddr");
+        cutoffBlock = (int)(values.get("atBlock"));
+        showAddr = (int)(values.get("withAddr"));
         detailed = values.get("detailed");
-        limit = values.get("limit");
+        limit = (int)(values.get("limit"));
 
         auto args = parser.args();
         for(size_t i=1; i<args.size(); ++i) {
